finatra: Fix SBT build for logging dependencies in test
Problem/Solution

Coursier is resolving multiple slf4j-api logging implementations from
its cache. We generally use the `slf4j-simple` logging implementation
for framework test logging, however the `inject-logback` project is
a specific Logback adaptation and needs to use `logback-classic` as a
test logging implementation. Unfortunately, coursier is now consistently
resolving `slf4j-simple` ahead of `logback` on the test classpath for
the `inject-logback` project making the tests fail.

Move around how we specify the `slf4j-simple` test logging dependency
to be defined explicitly where needed and with a `test-internal` scoping
which in theory tells ivy to not export it as a transitive dependency.
At least, that is what I interpret that means.

JIRA Issues: CSL-9265, CSL-9264

Differential Revision: https://phabricator.twitter.biz/D418520