inject-server: Fix issue in `c.t.inject.server.EmbeddedHttpClient`
Problem

The `c.t.inject.server.EmbeddedHttpClient` incorrectly disallows assertion
of empty response bodies via `withBody` and `withJsonBody`. This
prevents correctly asserting that a server is not yet healthy as the `/health`
endpoint returns an empty string. Thus even a not yet healthy server would
report as "healthy" to the testing infrastructure as long as the `health`
endpoint returns a `200 - OK` response.

Solution

Update the `c.t.inject.server.EmbeddedHttpClient` to allow for assertion of
an empty string as a response body. This means the sentinel value of `null`
alone indicates to perform no response body assertion. This was the previous
behavior before the logic was pulled into the `EmbeddedHttpClient`.

Add tests which correctly assert health of an underlying server during testing.

Result

It is possible to correctly assert an expected response body of an empty
string with the `c.t.inject.server.EmbeddedHttpClient`.

JIRA Issues: CSL-9300

Differential Revision: https://phabricator.twitter.biz/D422712