finatra: Exposing Listening Server's bound address in Thrift/HTTP server traits
== Problem ==

Currently the underlying `ListeningServer`'s in `ThriftServerTrait` and
`HttpServerTrait`  have a bound address which is not accessible outside of the
trait itself. This limits the implementations of these traits,  which need
access to them.

== Solution ==

Expose the underlying `ListeningServer`'s `boundAddress` via a `protected var`
in `ThriftServerTrait`. Also expose the http and https `boundAddress`'s via a
`protected var` in `ThriftServerTrait`.

== Result ==

Implementations of `ThriftServerTrait` and `HttpServerTrait` can access the
underlying `ListeningServer`'s bound address.

JIRA Issues: REVSRVG-20429

Differential Revision: https://phabricator.twitter.biz/D424745