finatra-kafka-streams: Track async call latencies with stat metrics
Problem

The latency of the resolution of Futures from the both the non-blocking
AsyncProcessor's processAsync call and AsyncTransformer's transformAsync call
together with the MaxOutstandingFuturesPerTask setting directly relate to how
often Kafka can commit offsets. Since there's currently no latency metric on
the time that a processAsync created Future takes to complete, it's hard to
tell if processAsync is taking longer than expected.

Solution

Add a latency histogram stat around the processAsync call so that users can
tell if processAsync is taking longer than expected and is thus causing the
Kafka commit frequency to be longer than expected.
Add a latency histogram stat around the transformAsync call so that users can
tell if transformAsync is taking longer than expected and is thus causing the
Kafka commit frequency to be longer than expected.

Results

A process_async_latency_ms metric will be available to expose quantiles of
AsyncProcessor.processAsync in milliseconds, and
a transform_async_latency_ms will be available to expose quantiles of
AsyncTransformer.transformAsync in milliseconds

JIRA Issues: VDI-2186

Differential Revision: https://phabricator.twitter.biz/D430688