finatra-kafkastreams: Add method to TopologyTesterTopic to write Kafka messages with custom headers to topics
Problem

The Finatra Kafka Streams topology integration test class does not provide a
way to write test messages with headers to a Kafka topic. Topologies which
need to test functionality based on Kafka message header values currently
would have to mock calls which access headers, requiring an internal
understanding of how headers are used by the library.

Solution

Added a new method to TopologyTesterTopic to allow callers to write messages
with a custom header to a test Kafka topic. The method simply uses the
existing overloaded create method on ConsumerRecordFactory with the passed in
header argument.

Result

It is easier and safer to test Kafka stream topologies with behavior based on
Kafka message headers (e.g., custom filters or transformations extending the
Processor API) without having to rely on mocking how Kafka message headers are
accessed.

Differential Revision: https://phabricator.twitter.biz/D424440