inject-app: Add `InjectorModule` to the TestInjector
Problem

The `c.t.inject.app.TestInjector` installs most of the same default
modules as `c.t.inject.app.App` (FlagsModule and TypeConvertersModule
through usage of `c.t.inject.app.Modules`).

However it does not install the `InjectorModule` which is done by App
as a framework module. To be consistent and have the same baseline of
framework modules, the TestInjector should also install this module by
default to give users the same base line object graph in both cases.

Solution

Add the InjectorModule to the TestInjector's list of required modules.

Differential Revision: https://phabricator.twitter.biz/D465943