inject-core: Change `c.t.inject.TwitterModuleFlags` to fail fast until parsed by default
Problem

Finatra provides an integration between an application's `c.t.app.Flags`
collection and Guice modules. Instances of a `c.t.app.Flag` added to the
application's `c.t.app.Flags` collection need to be parsed from command
line input in order to be fully processed.

As such, access to a `c.t.app.Flag` value should not be eager and must happen
no earlier than the `premain` phase of the application lifecycle.

`c.t.app.Flag` instances can be created with a "fail fast" behavior which was
added to aid in correct `c.t.app.Flag` value accessing. This behavior will throw
an exception if the `c.t.app.Flag` value is accessed before the instance has been
parsed. The default for this value is "false" for legacy reasons. Setting of the "fail fast"
behavior of a `c.t.app.Flag` is generally done by the application's `c.t.app.Flags`
collection since `c.t.app.Flag` are generally created by asking the `c.t.app.Flags`
collection for a new instance and thus is a configuration option on an application
which Finatra defaults to true. See `c.t.app.App#failfastOnFlagsNotParsed`.

However, when a `c.t.app.Flag` is created in a `TwitterModule` it is created
disjointly from the application's `c.t.app.Flags` collection and thus does not
inherit the application's setting for `failfastOnFlagsNotParsed` until the
`c.t.app.Flag` instance is added to the `c.t.app.Flags` in the `init` lifecycle phase.
This means it is possible to access the value of a `c.t.app.Flag` instance created within
a `TwitterModule` before parsing without failing fast.

Solution

Update to make `c.t.app.Flag` instances created within a `TwitterModule` have
the `failFastUntilParsed` value set to 'true' by default. While this value is
configurable inside a given `TwitterModule` it is STRONGLY encouraged that this
be the behavior at all times. We allow for toggling this functionality in order to
provide users an avenue for migration to the recommended way to define and
read the value of `c.t.app.Flag` instances.

Lastly, we reduce visibility on internal framework code. The public interface for
Modules in Finatra is the `c.t.inject.TwitterModule` itself and it should be used
directly instead of any traits that it is comprised from.

Result

Accessing the value a `c.t.app.Flag` instance created within a `c.t.inject.TwitterModule`
before the instance has been parsed will now result in an exception which typically will
fail server startup allowing users to catch issues with Flags in a `StartupTest` or a
`FeatureTest`.

JIRA Issues: CSL-9499

Differential Revision: https://phabricator.twitter.biz/D448047