inject-app: Remove finagle-core dependency
Problem

The libraries in finatra/inject are meant to only depend on each
other, 3rd party dependencies, or util. This breaks down for some of
the libraries but should hold true for the core inject libraries of
inject-core and inject-app. Currently, inject-app has a finagle-core
dependency solely because the `DtabResolution` utility trait is defined
in the library.

Solution

Move the `DtabResolution` utility trait to it's own library,
inject-dtab and remove the finagle-core dependency from
inject-app.

Result

A core finatra/inject library no longer has a finagle-core
dependency.

Differential Revision: https://phabricator.twitter.biz/D474298