inject-core: Deprecate c.t.inject.Mockito
Problem

With the availability of `mockito-scala` there is little
reason to have a Finatra specific bridge of the Specs2
MockitoSugar to work with ScalaTest. The `mockito-scala`
library obiviates the need to try to carry the Specs2
sugar to ScalaTest for Mockito as it provides a more
comprehensive Scala solution for use of Mockito in Scala
and with ScalaTest.

Solution

Mark `c.t.inject.Mockito` as deprecated.

JIRA Issues: CSL-9757

Differential Revision: https://phabricator.twitter.biz/D482531