finatra-http: Update HttpResponseFilter to allow extension
Problem/Solution

Allow extensions of the `c.t.finatra.http.filters.HttpResponseFilter`
to specify how to set the Location Header value into a Response.

Additionally, the `c.t.finatra.http.filters.HttpResponseFilter` should
not throw exceptions as it is installed "above" the `ExceptionMappingFilter`
and Http Response translation is not available. It is incorrect that
the inability to set a non-compliant 'Location' header value causes
an exception here, so the code has been updated to ensure non fatal
exceptions do not escape. Thus full qualification of a 'Location' header
value is best-effort.

JIRA Issues: CSL-9760

Differential Revision: https://phabricator.twitter.biz/D483793