inject-slf4j: Remove Jackson dependency
Problem/Solution

The finatra/inject/inject-slf4j library should not bring
in a jackson-annotations dependency as it is meant to be
a general library with minimal dependencies as it is used
in a lot of places. This dependency isn't needed here and
can be localized to a trait extension in finatra/jackson.

JIRA Issues: CSL-9793

Differential Revision: https://phabricator.twitter.biz/D487948