Use threadpool because producer.send() might block under certain cases.
Problem

Although claimed as aynchronous, the open source KafkaProducer send() might actually block if the metadata is not available or producer memory is full.
For that reason, we should execute send() in a threadpool instead to avoid it might block the finagle thread.

Solution

Execute producer.send() in thread pool and return the flatten future.

Result
Tried on all existing unitest and some of our application (like replicators.)

Differential Revision: https://phabricator.twitter.biz/D498808