finatra: improve `EmbeddedTwitterServer` startup time
Problem

Finatra's `EmbeddedTwitterServer` uses a polling
mechanism to determine whether the underlying
`twitterServer` is started and ready for testing.
There are 1 second polling intervals and there is
potential for race conditions that we have been
lucky to not observe (i.e. server is marked as started,
but ports are not exposed for a client to use).

Solution

Modify `inject.App` to emit lifecycle events and
`inject.TwitterServer` to mark the startup
completion event.

Have `EmbeddedTwitterServer` register observes
for lifecycle events and react to the emitted events.

Result

We no longer need to have custom logic for inject
vs non-inject `TwitterServer` startup. Test execution
is sped up for nearly all test targets using
`EmbeddedTwitterServer` and its inheritors.

JIRA Issues: CSL-9865

Differential Revision: https://phabricator.twitter.biz/D499999