[finatra-kafka] Add tracing information for topics
Problem
-------

Tracing information is absent from Kafka producers and consumers.

Solution
--------

Created `TracingKafkaConsumer` and `TracingKafkaProducer` to record topics in the trace if
actively tracing. The `TracingKafkaConsumer` additionally records the `TraceId` of the
producer that is present in the record's headers.

Result
------

The Kafka topic information is now available via traces.

JIRA Issues: REVSRE-7931

Differential Revision: https://phabricator.twitter.biz/D490711