finatra-inject: Bandaid racy AppendTestHttpServerFeatureTest
Problem

Now that the finatra feature tests are blazingly fast, it's possible for a
server to start up, and then see that all awaitables are finished, and shut down,
all before the first request has been sent.  This ruins this specific test, which
examines logged messages, and is surprised to see a logged message from the framework.

Solution

We're going to apply a bandaid of looking either for 5 or for 6. I'm not sure why the other
approach didn't work, but let's remove the KTF first and then work out the rest later.

Also remove the KTF annotation from the BUILD file.

JIRA Issues: CSL-9943

Differential Revision: https://phabricator.twitter.biz/D513544