finatra-jackson: Fix unknown properties handling.
Problem

'FAIL_ON_UNKNOWN_PROPERTIES' Deserialisation Feature doesn't work as
expected when the JSON contains less fields than the CaseClass.

In CaseClassDeserialiazer, the sequence of unknown fields is taken into
account only when the number of fields provided in the Json is greater
than the number of fields declared in the CaseClass.
For instance, when the case class contain some optional fields, the JSON
can contain less fields than the case class and one of them can be
unknown.

Solution

The solution is to always take into account unknown properties whatever
maybe the number of fields.

Result

When the JSON contains less fields overall than the CaseClass,
When the JSON provide one extra unknown field,
When the FAIL_ON_UNKNOWN_PROPERTIES is on,
Then an exception is thrown.