inject-server: EmbeddedTwitterServer to return `StartupTimeoutException`
Problem

We don't have a test for a server that fails to startup within the
`maxStartupTimeSeconds` for an EmbeddedTwitterServer. Adding this
test shows that the c.t.util.TimeoutException escapes instead of
a `StartupTimeoutException`.

Solution

Patch the EmbeddedTwitterServer await to catch the TimeoutException
and properly return a StartupTimeoutException with a properly
formatted error message.

Differential Revision: https://phabricator.twitter.biz/D519318