finatra-jackson: Fix issues with generic case class deserialization
Problem

Case class deserialization is not properly looking up the bound types
for a parameterized case class' generic types. This manifests in
different ways with one case being the incorrect deserialization of
a parameterized case class with an Option field of the type, e.g.,
```
case class B(value: String)
case class AGeneric[T](b: Option[T])
val aGeneric = mapper.parse[AGeneric[B]]("""{"b":{"value":"string"}}""")
```

This case is incorrectly deserialized as the `Option[T]` is interpreted
as a Map instead of as an Option.

Solution

First make sure we are properly handling Option types when determining
if we need to create a fully specified JavaType with generic parameters. We
were previously skipping all Options in this check which is not correct as
the Option field type may be genericized to a case class type parameter.

Secondly, re-work the Jackson `Types.javaType` which converts from the
ScalaType to a JavaType for deserialization to properly look up the bound
types for a given field's generic types from the case class bound parameter
types in order to properly deserialize the JSON into the correct parameterized
type.

Result

Fixes #547. Better handling of generic case class deserialization.

JIRA Issues: CSL-10088

Differential Revision: https://phabricator.twitter.biz/D532768