inject-server: Awaiter.any don't block on empty list
Problem/Solution

c.t.inject.server.Awaiter.any will incorrectly try to await on an
empty list of Awaitables causing the code to hang indefinitely. Let's
short-circuit to not attempt the logic on an empty list. Additionally,
the return type of a `Future[Unit]` doesn't match the intent of the
awaiting and should mirror `Await.ready` in returning a Unit thus we
`Await` on the TimerTask.close call to ensure we get to a Unit return.

Differential Revision: https://phabricator.twitter.biz/D537727