finatra-http - fix HttpWarmer's use of call-by-name which can cause subtle bugs
Problem
-------

HttpWarmer.send uses call-by-name for the request parameter, which in certain
cases can cause the desired request object to be recreated multiple times within
the method. This prevents the mutation that the header method is trying to do.

Solution
--------

HttpWarmer.send resolves the request once per loop, which fixes the mutation
issue.

Result
------

HttpWarmer should work correctly for code passing in a supplying function rather
than a variable on this parameter.

JIRA Issues: CSL-10154

Differential Revision: https://phabricator.twitter.biz/D547310