finatra-validation: Fix Validator for Case Classes with Other Fields
Problem

The `Validator` throws an `IndexOutOfBoundsException` when trying to
validate a case class which contains fields that are not set via
constructor parameters. This is because of `Validator`'s use of
`AnnotationUtils`, and that in this scenario, `Validator` is breaking
the contract of how it should be used. `Validator` passed in an array
of fields returned from `getDeclaredFields`, which will return ALL
fields. `AnnotationUtils` is only expecting a subset of the list of
parameters that it would find via the constructor. The two don't match
and so an exception occurs.

Solution

After consultation with the team, we do not want to change
`AnnotationUtils` at this time. So, we instead fix `Validator` to abide
by the contract of the method and pass in the appropriate constructor
parameters, rather than ALL fields. So we swap out the use of
`getDeclaredFields` with a retrieval of parameters based on the first
constructor.

JIRA Issues: CSL-10170

Differential Revision: https://phabricator.twitter.biz/D549253