finatra: Fix up TwitterServer examples
Problem/Solution

The pubsub example code used for the injectable TwitterServer examples
in both Scala and Java can use some work. The Publishers and Subscribers
define Future pools but don't use them consistently and can be simplified.

The feature tests aren't written to be able to expect behavior on the
subscriber reads.

Update the code and tests to flesh out the examples and make them a bit
less contrived.

Differential Revision: https://phabricator.twitter.biz/D546475