inject-thrift-client: Fix DarkTrafficFilter Lookup for Inherited Methods
Problem

When using `ReqRepServicePerEndpoint` with Finatra's
`DarkTrafficFilter`, the `DarkTrafficFilter` is incapable of correctly
looking up an inherited Thrift method. This occurs when a Thrift
'service' extends another 'service'. The method ends up being defined
on the *base* service and not the *leaf* service. The
`DarkTrafficFilter` uses `getDeclaredMethod` which will only find
methods declared on the *leaf* service.

Solution

Change the use of `getDeclaredMethod` to `getMethod`. Using `getMethod`
allows the `DarkTrafficFilter` to find a method on the *base* service.

Result

The `DarkTrafficFilter` works properly for inherited methods when
'lookupByMethod' is set to true.

JIRA Issues: CSL-10209

Differential Revision: https://phabricator.twitter.biz/D553361