[Finatra]: Adding support for passing chain in the TLS sever trait
Problem
-------

There is currently no way to provide a chain in the TLS server trait exposed by Finatra
for configuring a HTTPS server, if you are using a custom chain. This leads to the onus
on the developer to re-implement the same interface with a chain flag and pass it in the
SslServerConfiguration.

Solution
--------

Adding a `chain.flag` path in TLS server trait and using it optionally for creating a
CertKeyAndChain object to pass it to the SslServerConfiguration. Also diabling the
defaultHttpPort by default so that server only serves on the HTTPS port.

JIRA Issues: PSEC-8692

Differential Revision: https://phabricator.twitter.biz/D553718