utils: Re-vamp ClassUtils.isCaseClass and add ClassUtilsTest
Problem/Solution

Let's use the TypesApi for help in determining if a class is
a case class and let's also add some tests.

Additionally, `class.getSimpleName` can throw an `InternalError`
for malformed package/class names. Let's handle this case in
`ClassUtils.simpleName` such that the method is a best effort at
producing a simple class name by parsing the last token of the fully
qualified class name after replacing all `$` with `.` when the call to
`clazz.simpleName` fails. We also add tests here.

Differential Revision: https://phabricator.twitter.biz/D566069