utils: Undo switch to TypesApi for ClassUtils as this fails for some cases in 2.11
Problem

Using the TypesApi to help determine if a class is a case class
fails for some supported cases in 2.11. Let's revert to the more
straightforward way previously done which supports the generic
case class usages in Scala 2.11.

Solution

Put back the old way for Scala 2.11 support.

Result

We at least have unit tests for this code now.

Differential Revision: https://phabricator.twitter.biz/D566596