finatra: Add some help for those with no Jackson familiarity
Problem/Solution

Our Jackson documentation assumes a level of familiarity with Jackson,
databinding, and the ObjectMapper that some users may not have. We
also don't have any examples of basic usage for serializing or
deserializing case classes with the Finatra `ScalaObjectMapper`.

Let's add some caveats and pointers to available tutorials for
helping users find basic information on Jackson. And let's also add a
few simple examples of serialization and deserialization with the
`ScalaObjectMapper` while also pointing users to the tests which have
numerous more examples.

Result

Hopefully users with little to no understanding of Jackson are able
to glean useful information to help get them started.

Differential Revision: https://phabricator.twitter.biz/D567220