inject-thrift-client: Add tests to cover legacy behavior in DarkTrafficFilter
Problem

In investigating DarkTrafficFilter's incapability of correctly looking up an
inherited Thrift method, it was found that a branch of DarkTrafficFilter's
`getService` method did not have any tests; specifically, the branch when
`lookupByMethod` was false.

Solution

Share the tests that test the inherited services with the DarkTrafficFilter, and
test the branch of logic in `getService` where `lookupByMethod` is false by
creating the servers using the legacy API (`ServiceIface`).

Result

The DarkTrafficFilter works properly for inherited methods when 'lookupByMethod'
is set to false.

JIRA Issues: CSL-10214

Differential Revision: https://phabricator.twitter.biz/D569586