finatra-http: Split CallbackConverter Interface from Implementation
Problem

At the heart of the `CallbackConverter` is a function which transforms
a callback from a `RequestType` to a `ResponseType` to one that works
on a finagle-http `Request` and returns a `Future` of a finagle-http
`Response`. The implementation of this function is quite complex and
heavyweight. This makes testing individual pieces of finatra-http's
routing library challenging, as a full on `CallbackConverter`
implementation is required to construct a `Route`, even when the desire
is to only test the construction of routes from a single controller.

Solution

Split the `CallbackConverter`'s interface away from its implementation,
so that a simplified version of the `CallbackConverter` can be used for
testing instead, when actual responses are not being tested.

JIRA Issues: CSL-10336

Differential Revision: https://phabricator.twitter.biz/D573997