finatra-{httpclient, jackson,http}: Cross-build for Scala 2.13
Problem

httpclient is not cross-built for Scala 2.13.
jackson is not cross-built for Scala 2.13.
http is not cross-built for Scala 2.13.

Solution

Update httpclient module to cross-build for Scala 2.13 using new SBT
settings.
Update jackson module to cross-build for Scala 2.13 using new SBT
settings.
Update http module to cross-build for Scala 2.13 using new SBT
settings.

Result

httpclient is cross-built for Scala 2.13.
jackson is cross-built for Scala 2.13.
http is cross-built for Scala 2.13.

Closes #554

Signed-off-by: Yufan Gong <yufang@twitter.com>

Differential Revision: https://phabricator.twitter.biz/D574391