finatra/inject: Simpler customizing timeout for inMemoryStats
Problem:

The only way to set a custom timeout for a metric is with a predicate
`T => Boolean`. If users want to check whether a metric equals a specific
value, it is simpler to pass the expected value instead of wrapping it in a
predicate like `(_ = 1.0f)`.

Solution:

Add a new, non-predicate version of waitFor that waits for an expected
value within a specific timeout. This non-dynamic case keeps waitFor consistent
with other methods that assert or validate expected values.

JIRA Issues: CSL-9951

Differential Revision: https://phabricator.twitter.biz/D576147