finatra-kafka-streams: Add PersistentTimerValueStore
Problem

The `PersistentTimerStore` does not allow storing a value. If you encode the value into the key,
then if the value's serialization changes (due to a thrift change, perhaps) then events in the timer
store will never be deleted.

Solution

Create a `PersistentTimerValueStore` which stores a value along with the key into the store.
Also change `PersistentTimerStore` to be an extension of `PersistentTimerValueStore` which
doesn't store any value.

JIRA Issues: DSTR-8295, ADAT-5060

Differential Revision: https://phabricator.twitter.biz/D583020