finatra: Fix the StackClientModuleTraitTest for sbt build
Problem

In StackClientModuleTraitTest we're checking that the stats receiver is
the NullStatsReceiver but this is no longer the case in the OSS build
since finagle-stats is now on the classpath and that has a service loaded
stats receiver.

Solution

We don't actually care whether the stats receiver is null or not, just
that it's injected. Therefore, just check that it's not null. Note that
this is redundant since it's a required component of creating the client
but we check it anyway just to be explicit.

JIRA Issues: CSL-10435

Differential Revision: https://phabricator.twitter.biz/D590339