[3rdparty] mark 3rdparty dependencies not in 3rdparty as intransitive = False
Problem / Solution

We'd like to move a world where all dependencies are explicitly mentioned in BUILD files. Make a small amount of progress by marking current policy-violating dependencies as `intransitive = False`. This is a pre-req to changing the default which is a pre-req to changing all deps.

JIRA Issues: DEAD-38

Differential Revision: https://phabricator.twitter.biz/D597611