finatra-kafka-streams: Fix delayWithStore DSL call
Problem

The timer store created by the `delayWithStore` DSL call was named
based on the hashcode of the implicit class, but that lead to different
shards using different timer stores, which would lead to orphan stores
when a shard restarts or on a rebalance.

Solution

Add a `storeKey` flag to the `delayWithStore` DSL call so that the
caller can specify a unique string which will allow the topology
to have multiple `delayWithStore` calls with different stores,
while using the same store across shards.

JIRA Issues: DSTR-8047

Differential Revision: https://phabricator.twitter.biz/D593593