[Bazel] Export Pants graph to Bazel using bazel-multiversion
This is a rework of {D563922} that exports Pants graph to Bazel.

The main change is that this adds `3rdparty/jvm_deps.bzl` generated using a CLI called multiversion (https://github.com/twitter/bazel-multiversion), which in turns calls back to pants as `./pants bazel-multideps <3rdparty target>` to generate a JSON export of Pants graph.

The multiversion CLI then generates a cohesive that implements unified and multiversion resolution of the entire graph. To inform this tool, we will now encode `version_scheme` information in the BUILD file. For example, Kafka will use [pvp](https://pvp.haskell.org/), which will differentiate Kafka 2.2 from Kafka 2.5.

The generated `3rdparty/jvm_deps.bzl` contains `scala_import` rules for all of the JARs in the graph, as well as the dependency relationship between them. Translation Layer is updated to use those nodes.

JIRA Issues: DPB-13042

Differential Revision: https://phabricator.twitter.biz/D597747