inject-thrift-client: Update MultiServerDarkTrafficFilterFeatureTest to wait for metric value
Problem

The MultiServerDarkTrafficFilterFeatureTest can be flaky due to the racing
of metrics appearing in the sink and querying for said metrics.

Solution

Update the metric assert to use `waitFor` to poll for a small amount of time
to allow for the metric to appear in the sink.

Result

Less chance of the metric attempting to be read before it is in the sink.

JIRA Issues: CSL-10636

Differential Revision: https://phabricator.twitter.biz/D614143