finatra: rename finatra/http project to finatra/http-server
Problem/Solution

The finatra/http project encapsulates the implementation of
the `HttpServer` and its surrounding utilities. There are
components in this project that we would also like to be
shared. By renaming the project to finatra/http-server, it
will allow us to extract those shared components to a
finatra/http project later.

JIRA Issues: CSL-10652

Differential Revision: https://phabricator.twitter.biz/D616257