finatra: Rename TwitterModule.closeOnExit to TwitterModule.onExit
Problem

Finatra's `TwitterModule.closeOnExit` suggests that it takes a `Closable` as an argument, just
like like `App.closeOnExit`. This, in fact, is not true and the TwitterModule-version of `closeOnExit`
takes a thunk, `() => Unit`. This causes both confusion with users and resource-leaks when the
clean-up code was copying from `App` to a `TwitterModule` (it now does nothing but happily compiles).

Solution

Rename `TwitterModule.closeOnExit` to `onExit`, a symetrical variant of this API within `App`.

JIRA Issues: CSL-10271

Differential Revision: https://phabricator.twitter.biz/D621095