finatra: move internal/marshalling to http-core
Problem:
we want to decouple the message body components from http-server.

Solution:
Move `internal/marshalling` to `http-core`'s marshalling package. Also
move `MessageBodyFlagsModule` to `http-core` as it's a dependency
on the writer default implementation.

Result:
We no longer need to `internal` package namespace as we can make the
default reader and writer implementations private to the marshalling
package.

Introduce the `MessageBodyManagerModule` which provides these
implementations. The `MessageBodyModule` in http-server can include this
as a dependency to have the readers and writers present in the object graph.

JIRA Issues: CSL-10731

Differential Revision: https://phabricator.twitter.biz/D626600