finatra: Singleton MessageBodyManager configured by MessageBodyModule
Problem
We want to decouple MessageBodyManager from http-server and use a
consistent pattern in configuration.

Solution
First Step: Make `MessageBodyManager` a singleton, configurable via the builder
pattern exposed in the `MessageBodyModule`.

Result
Since the `HttpRouter` mutates the `MessageBodyManager`, the mutable API
must be kept until the `HttpRouter` is updated to make use of the builder. For
this reason the `ConcurrentHashMap`s in `MessageBodyManager` are kept.

JIRA Issues: CSL-10660

Differential Revision: https://phabricator.twitter.biz/D621755