finatra-kafka-streams: Update AsyncTransformer to support threadpools
Problem

The AsyncTransformer should support using a threadpool for Kafka Streams
workloads that are CPU-bound and need to make use of additional threads for
event processing. When using the AsyncTransformer with a threadpool, the
respond in `addFuture` executes in a different thread than the one calling
`onFlush`, so there is a race condition where sometimes not all permits have
been released by the time the assert statement
`assert(numOutstandingFutures ==  0)` is evaluated in `onFlush`.

Solution

This diff changes the respond to a transform in order to ensure that all
permits get released and `asyncFailure` gets set (if needed) before the Future
is satisfied. In order to keep the semantics of transform equivalent to respond
(with the exception of ensuring the callback completes), return a Future.value
if the Future is successful and a Future.exception in the case of a failed
Future. The returned values are needed because calling `.unit` on a Future
discards the result, but keeps failed Futures as is.

Result

The AsyncTransformer can be used with a threadpool, thereby allowing CPU-bound
Kafka Streams workloads to use additional threads for event processing.

JIRA Issues: PUBSUB-76829

Differential Revision: https://phabricator.twitter.biz/D611608