Finatra Kafka: Set the default linger to 5 milliseconds.
Problem

The current linger is 0 ms, which is not ideal for low throughput topic and batching.

Solution

Set linger default to 5 milliseconds, which will improve batching, reduce request volume (and thus improve the tail latency).

Result

Improve the batching, tail latency (less number of requests)  and with sacrifice on direct produce latency.

Differential Revision: https://phabricator.twitter.biz/D652832