finatra-jackson: Fix test to deal with Scala 2.13
Problem

In Scala 2.13.1 the message for the returned java.util.NoSuchElementException
is null for a HashMap look up whereas in Scala 2.12.12 it is non null (and
useful). We pass the message on as part of case class deserialization and
expect behavior against it in a test.

Solution

Update the test accordingly to handle the change in situation in Scala 2.13.1.

Differential Revision: https://phabricator.twitter.biz/D666142