[Bazel][finatra] Fix conflicts between `jvm_binary`'s `basename` and `*_library`'s `name`.
Problem/Solution

Before the change, two targets produced a jar with the same name; which is problematic as it means that compiling targets in a different order could lead to a different outcome.

Remove unnecessary target aliases in favour of explicitly listing the full path to the dependencies to address the issue.

JIRA Issues: DPB-13191

Differential Revision: https://phabricator.twitter.biz/D649045