[finatra/kafka] Deprecate TracingKafkaConsumer and provide alternative
**Problem**

Currently, Kafka consumers generally produce single-span traces because the current implementation
relies on the caller of the `poll` operation to set up tracing. After doing some evaluation it turns
out that we need to change the entire API for the consumer for tracing information to be propagated.

**Solution**

Deprecated the `TracingKafkaConsumer` client which extends the `KafkaConsumer` as it will only create
single-span traces. We now provide `KafkaConsumerTracer` which has a `trace` method that accepts the
consumed records and sets up tracing for the consumer and any future RPC calls by accepting a
continuation function in the API which will have tracing set up appropriately. We also provide
`KafkaProducerTraceAnnotator` and `KafkaConsumerTraceAnnotator` services to allow users to add
custom trace annotations to producer and consumer spans.

**Result**

The alternative `trace` method will enable users to perform more robust tracing.

Differential Revision: https://phabricator.twitter.biz/D649655