finatra/jackson: Defer field node validation until after deserializer resolution
Problem

We currently do not allow for a custom deserializer specified for a field to
allow parsing of `null` JSON values into `null` types during deserialization
as we assert that the field JsonNode cannot be null and cannot have `null`
value. This prevents a user from defining and using a custom deserializer which
allows the field to be parsed into a `null` type.

Based on the PR conversation here: https://github.com/twitter/finatra/pull/562

Additionally, there is a bug in how String case class fields are handled when the
incoming JSON is not a String-type. The current code incorrectly returns an empty
string when the JSON field value is parsed into Jackson `ContainerNode` or `ObjectNode`
types and an incorrect `toString` representation for a `PojoNode` type.

Solution

Refactor the code to proxy responsibility for validating the field JsonNode
to the invoked deserializer and not impose the constraints of the `CaseClassDeserializer`
onto it.

We also now correctly represent a non JSON field value as a string in cases where it is
being deserialized into the case class field of type String.

Result

Users have more control when defining a custom deserializer on how it deserializes
the field into a type.

No more empty strings returned for some deserialized JSON field values.

JIRA Issues: CSL-10988, CSL-10994

Differential Revision: https://phabricator.twitter.biz/D676938